+++ harper-cli/Cargo.toml
@@ -12,8 +12,9 @@ ariadne = "0.4.1"
 clap = { version = "4.5.23", features = ["derive"] }
 harper-core = { path = "../harper-core", version = "0.14.0" }
 harper-comments = { path = "../harper-comments", version = "0.14.0" }
+harper-typst = { path = "../harper-typst", version = "0.14.0", optional = true }
 serde_json = "1.0.133"
 
 [features]
 default = ["typst"]
-typst = ["harper-core/typst"]
+typst = ["dep:harper-typst"]
