+++ harper-cli/src/main.rs
@@ -172,7 +172,7 @@ fn load_file(file: &Path) -> anyhow::Result<(Document, String)> {
         match file.extension().map(|v| v.to_str().unwrap()) {
             Some("md") => Box::new(Markdown),
             #[cfg(feature = "typst")]
-            Some("typ") => Box::new(harper_core::parsers::Typst),
+            Some("typ") => Box::new(harper_typst::Typst),
             _ => Box::new(
                 CommentParser::new_from_filename(file)
                     .map(Box::new)
