+++ harper-comments/Cargo.toml
@@ -32,6 +32,3 @@ tree-sitter-haskell = "0.15.0"
 
 [dev-dependencies]
 paste = "1.0.15"
-
-[features]
-typst = ["harper-core/typst"]
