+++ harper-core/Cargo.toml
@@ -24,7 +24,6 @@ thiserror = "2.0.9"
 unicode-blocks = "0.1.9"
 unicode-width = "0.2.0"
 levenshtein_automata = { version = "0.2.1", features = ["fst_automaton"] }
-typst-syntax = { version = "0.12.0", optional = true }
 
 [dev-dependencies]
 criterion = { version = "0.5.1", default-features = false }
@@ -35,5 +34,5 @@ name = "parse_demo"
 harness = false
 
 [features]
-typst = ["dep:typst-syntax"]
+default = []
 concurrent = []
