+++ harper-html/Cargo.toml
@@ -14,6 +14,3 @@ tree-sitter = "0.20.10"
 
 [dev-dependencies]
 paste = "1.0.15"
-
-[features]
-typst = ["harper-core/typst"]
