+++ harper-ls/Cargo.toml
@@ -10,6 +10,7 @@ repository = "https://github.com/automattic/harper"
 [dependencies]
 harper-core = { path = "../harper-core", version = "0.14.0", features = ["concurrent"] }
 harper-comments = { path = "../harper-comments", version = "0.14.0" }
+harper-typst = { path = "../harper-typst", version = "0.14.0", optional = true }
 harper-html = { path = "../harper-html", version = "0.14.0" }
 tower-lsp = "0.20.0"
 tokio = { version = "1.42.0", features = ["fs", "rt", "rt-multi-thread", "macros", "io-std", "io-util", "net"] }
@@ -28,4 +29,4 @@ serde = { version = "1.0.214", features = ["derive"] }
 
 [features]
 default = ["typst"]
-typst = ["harper-core/typst"]
+typst = ["dep:harper-typst"]
