+++ harper-tree-sitter/Cargo.toml
@@ -9,7 +9,3 @@ repository = "https://github.com/automattic/harper"
 [dependencies]
 harper-core = { path = "../harper-core", version = "0.14.0" }
 tree-sitter = "0.20.10"
-
-[features]
-default = ["typst"]
-typst = ["harper-core/typst"]
