+++ harper-typst/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "harper-typst"
+version = "0.14.0"
+edition = "2021"
+
+[dependencies]
+harper-core = { path = "../harper-core", version = "0.14.0" }
+typst-syntax = { version = "0.12.0" }
+ordered-float = { version = "4.6.0", features = ["serde"] }
+itertools = "0.13.0"
+paste = "1.0.14"
